Week07/requirements.txt
datasets
git+https://github.com/openai/CLIP.git
torch
torchvision
numpy
Pillow
scikit-learn
matplotlib
